requirements.txt
fastapi==0.128.0
pydantic==2.12.5
python-dotenv==1.2.1
SQLAlchemy==2.0.46
uvicorn[standard]==0.30.0
pymysql==1.1.0

